vendor/bflb: Emulate delay functions by nanosleep
BFLB SDK use bflb_mtimer_delay_us/bflb_mtimer_delay_ms function to delay,
but it depends on the system's mtimer hardware, which is occupied by OS
kernel, so it is not suitable for Wasm app.

This patch adds delay functions based on nanosleep, which is a standard C
library function and does not depend on any hardware. This makes it suitable
for Wasm app.

Signed-off-by: Huang Qi <huangqi3@xiaomi.com>